Lesson-7: review fixes
