+++ Project/api/helpers/DBconnection.js
@@ -1,5 +1,5 @@
 import * as mysql from 'mysql2';
-import { DB_ACCESS } from './config.js';
+import { DB_ACCESS } from '../../config.js';
 
 export function createConnection() {
   const connection = mysql.createConnection({
