+++ Project/api/helpers/models/patientModel.js
@@ -1,5 +1,5 @@
 import { promisify } from 'util';
-import { createConnection } from '../../DBconnection.js';
+import { createConnection } from '../DBconnection.js';
 
 export async function patientDefine() {
   const connection = createConnection();
