+++ Project/api/helpers/schemas/AddPatientSchema.js
@@ -1,7 +1,8 @@
-import { MAX_LENGTH, MIN_LENGTH } from '../../constants.js';
+import { MAX_LENGTH, MIN_LENGTH } from '../../../constants.js';
 
 export const AddPatientSchema = {
   type: 'string',
   maxLength: MAX_LENGTH,
   minLength: MIN_LENGTH,
+  pattern: '[a-zA-Z]+',
 };
